[Silver II] Title: 행사 준비, Time: 624 ms, Memory: 54796 KB -BaekjoonHub

+++ 백준/Silver/30022. 행사 준비/행사 준비.java
@@ -4,9 +4,9 @@ import java.util.Arrays;
 
 public class Main {
 	static class Item implements Comparable<Item> {
-		long a, b, diff;
+		int a, b, diff;
 
-		public Item(long a, long b) {
+		public Item(int a, int b) {
 			super();
 			this.a = a;
 			this.b = b;
